RDRP-449: period year function moved and columns confirmed
