+++ src/developer_config.yaml
@@ -5,7 +5,7 @@ global:
   postcode_csv_check: False
   load_historic_data: False
   load_manual_outliers: False
-  output_full_responses: True
+  output_full_responses: False
   output_outlier_qa : False
   output_estimation_qa: False
   output_imputation_qa: False
