+++ src/outputs/long_form.py
@@ -7,69 +7,10 @@ from typing import Callable, Dict, Any
 
 import src.outputs.map_output_cols as map_o
 from src.staging.validation import load_schema
-from src.outputs.outputs_helpers import create_output_df
+from src.outputs.outputs_helpers import create_output_df, create_period_year
 
 
 OutputMainLogger = logging.getLogger(__name__)
-
-
-def create_period_year(df: pd.DataFrame) -> pd.DataFrame:
-    """Created year column for long form output
-
-    The 'period_year' column is added containing the year in form 'YYYY'.
-
-    Args:
-        df (pd.DataFrame): The main dataframe to be used for long form output.
-
-    Returns:
-        pd.DataFrame: returns long form output data frame with added new col
-    """
-
-    # Extracted the year from period and crated new columns 'period_year'
-    df["period_year"] = df["period"].astype("str").str[:4]
-
-    return df
-
-
-def create_headcount_cols(
-    df: pd.DataFrame,
-    round_val: int = 4,
-) -> pd.DataFrame:
-    """Create new columns with headcounts for civil and defence.
-
-    Column '705' contains the total headcount value, and
-    from this the headcount values for civil and defence are calculated
-    based on the percentages of civil and defence in columns '706' (civil)
-    and '707' (defence). Note that columns '706' and '707' measure different
-    things to '705' so will not in general total to the '705' value.
-
-    Args:
-        df (pd.DataFrame): The survey dataframe being prepared for
-            long form output.
-        round_val (int): The number of decimal places for rounding.
-
-    Returns:
-        pd.DataFrame: The dataframe with extra columns for civil and
-            defence headcount values.
-    """
-    df = df.copy()
-    # Use np.where to avoid division by zero.
-    df["headcount_civil"] = np.where(
-        df["706"] + df["707"] != 0,  # noqa
-        df["705"] * df["706"] / (df["706"] + df["707"]),
-        0,
-    )
-
-    df["headcount_defence"] = np.where(
-        df["706"] + df["707"] != 0,  # noqa
-        df["705"] * df["707"] / (df["706"] + df["707"]),
-        0,
-    )
-
-    df["headcount_civil"] = round(df["headcount_civil"], round_val)
-    df["headcount_defence"] = round(df["headcount_defence"], round_val)
-
-    return df
 
 
 def run_longform_prep(
@@ -97,9 +38,6 @@ def run_longform_prep(
         & (df["form_status"].isin(["600", "800"]))
         & (df["instance"] == 0)
     ]
-
-    # create columns for headcounts for civil and defense
-    df = create_headcount_cols(df, round_val)
 
     return df
 
@@ -136,18 +74,6 @@ def output_long_form(
 
     # Join foriegn ownership column using ultfoc mapper
     df = map_o.join_fgn_ownership(df, ultfoc_mapper)
-
-    # Map to the CORA statuses from the statusencoded column
-    df = map_o.create_cora_status_col(df, cora_mapper)
-
-    # Map the sizebands based on frozen employment
-    df = map_o.map_sizebands(df)
-
-    # Map the itl regions using the postcodes
-    df = map_o.join_itl_regions(df, postcode_itl_mapper)
-
-    # Map q713 and q714 to numeric format
-    df = map_o.map_to_numeric(df)
 
     # Prepare the longform output dataframe
     longform_output = run_longform_prep(df, round_val=4)
