+++ src/outputs/outputs_helpers.py
@@ -34,3 +34,21 @@ def create_output_df(df: pd.DataFrame, output_schema: dict) -> pd.DataFrame:
     output_df = output_df[colname_dict.values()]
 
     return output_df
+
+
+def create_period_year(df: pd.DataFrame) -> pd.DataFrame:
+    """Created year column for short form output
+
+    The 'period_year' column is added containing the year in form 'YYYY'.
+
+    Args:
+        df (pd.DataFrame): The main dataframe to be used for short form output.
+
+    Returns:
+        pd.DataFrame: returns short form output data frame with added new col
+    """
+
+    # Extracted the year from period and crated new columns 'period_year'
+    df["period_year"] = df["period"].astype("str").str[:4]
+
+    return df
