+++ src/outputs/short_form.py
@@ -7,28 +7,10 @@ from typing import Callable, Dict, Any
 
 import src.outputs.map_output_cols as map_o
 from src.staging.validation import load_schema
-from src.outputs.outputs_helpers import create_output_df
+from src.outputs.outputs_helpers import create_output_df, create_period_year
 
 
 OutputMainLogger = logging.getLogger(__name__)
-
-
-def create_period_year(df: pd.DataFrame) -> pd.DataFrame:
-    """Created year column for short form output
-
-    The 'period_year' column is added containing the year in form 'YYYY'.
-
-    Args:
-        df (pd.DataFrame): The main dataframe to be used for short form output.
-
-    Returns:
-        pd.DataFrame: returns short form output data frame with added new col
-    """
-
-    # Extracted the year from period and crated new columns 'period_year'
-    df["period_year"] = df["period"].astype("str").str[:4]
-
-    return df
 
 
 def create_headcount_cols(
